Pin distribute to 0.6.34.
For some reason the buildout that we create in the test run gets
updated to a newer version of distribute, but not the newest.  So we
pin that version here, otherwise the first doctest goes wrong
because you get a few extra lines at the start, about upgrading and
restarting the buildout.