Accept the comment if it contains whitespace
Currently, the value is simply split by whitespace; thus, the "package"
cannot contain blanks, and we can't tell how many of the following words
are part of the comment.

If the zcml value one day would be split by lines only, we would be able
to safely skip the end of the line, following the '#'.

If the comment doesn't contain blanks, we can't know this;
but the only way for a comment to contain blanks is splitting by lines.
So we can safely accept the '#text with blanks' or '# text with blanks'
"package" as a comment and skip it.