Restored ability to use own explicit version of zodb-temporary-storage.
You used to be able to set `zodb-temporary-storage = <some config />` and have this config in the `site.zcml`.
Most important use of this was to let this empty, effectively disabling the temporary storage.
PR https://github.com/plone/plone.recipe.zope2instance/pull/93 changed this so you could do this in a more natural way.
by setting `zodb-temporary-storage = false`.

But this (accidentally) removed the possibility for setting an explicit own version of the temporary storage snippet,
and meant the only options now were false or true.
The documentation still mentions:

   If given Zope's default temporary storage definition will be replaced by
   the lines of this parameter.